requirements.txt
flask==2.2.1
keras==2.9.0
numpy==1.22.4
opencv-python==4.6.0
pillow==9.2.0
tensorflow==2.5.0
tensorflow-addons==0.17.1